change return type into double instead of int

+++ src/main/java/rateexchangeapp/Currency.java
@@ -2,11 +2,10 @@ package rateexchangeapp;
 
 import io.temporal.activity.ActivityInterface;
 import io.temporal.activity.ActivityMethod;
-import org.json.JSONObject;
 
 @ActivityInterface
 public interface Currency {
 
     @ActivityMethod
-    int exchangeCurrencyData();
+    double exchangeCurrencyData();
 }
